% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing_nodes.R
\name{add_missing_nodes}
\alias{add_missing_nodes}
\title{add_missing_nodes}
\usage{
add_missing_nodes(p1, p2, ...)
}
\arguments{
\item{p1}{- partition 1 - named vector}

\item{p2}{- partition 2 - named vector}
}
\value{
none
}
\description{
add missing nodes to a partition
}
\examples{
add_missing_nodes(p1, p2)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
