% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_plot.R
\name{annotate_plot}
\alias{annotate_plot}
\title{annotate_plot}
\usage{
annotate_plot(plot, stats)
}
\arguments{
\item{plot}{- ggplot}
}
\value{
none
}
\description{
Function to add significance stars given a protein boxplot and stats.
}
\examples{
annotate_plot(plot, stats)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
