% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_stars.R
\name{annotate_stars}
\alias{annotate_stars}
\title{annotate_stars}
\usage{
annotate_stars(plot, stats)
}
\arguments{
\item{plot}{- ggplot}
}
\value{
none
}
\description{
Function to add significance stars given a protein boxplot,
stats with FDR column and the column to be labeled.
}
\examples{
annotate_stars(plot, stats)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
