% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateTopTags.R
\name{annotateTopTags}
\alias{annotateTopTags}
\title{annotateTopTags}
\usage{
annotateTopTags(y_TT)
}
\arguments{
\item{y_TT}{}
}
\value{
none
}
\description{
A function to annotate DAP candidates.
}
\examples{
annotateTopTags(y_TT)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
