% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCytoscapeGraph.R
\name{createCytoscapeGraph}
\alias{createCytoscapeGraph}
\title{createCytoscapeGraph}
\usage{
createCytoscapeGraph(
  g0,
  g1,
  tissue,
  nodes = NULL,
  n_cutoffs = 1000,
  netwdir = getwd(),
  netw_layout = "force-directed",
  t_sleep = 2,
  min_edge_transparency = 255,
  max_edge_transparency = 255,
  min_node_size = 35,
  max_node_size = 100,
  min_edge_color = col2hex("gray"),
  max_edge_color = sizzling_red,
  threshold = NULL
)
}
\description{
createCytoscapeGraph
}
