% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deparseLayout.R
\name{deparseLayout}
\alias{deparseLayout}
\title{deparseLayout}
\usage{
deparseLayout(layout.name, layout.properties)
}
\value{
None
}
\description{
Deparse RCy3 layout arguuments to be pased to LayoutNetwork()
}
\examples{
# Create a graph layout to be applied to the layers.
# To see all layout properties:
getLayoutPropertyNames("force-directed")
layout.name <- "force-directed"
layout.properties <- list(
  defaultEdgeWeight = 0,
  defaultSpringCoefficient = 0.0001,
  edgeAttribute = "weight",
  defaultNodeMass = 10,
  defaultSpringLength = 1,
  numIterations = 100
)
myLayout <- deparseLayout(layout.name, layout.properties)
}
\references{
\url{}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{list}
\keyword{named}
