% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_down.R
\name{fill_down}
\alias{fill_down}
\title{fill down}
\usage{
fill_down(x, blank = is.na)
}
\arguments{
\item{x}{column vector with blank values.}

\item{blank}{logic vector specifying blank values.}
}
\value{
None
}
\description{
Fill a data frame with missing values.
Missing values are replaced with the value above them in a column.
}
\examples{
fill_down()
}
\references{
\url{https://stackoverflow.com/questions/10554741/fill-in-data-frame-with-values-from-rows-above}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{blank}
\keyword{down}
\keyword{fill}
\keyword{missing}
\keyword{values}
