% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_modules.R
\name{filter_modules}
\alias{filter_modules}
\title{filter_modules}
\usage{
filter_modules(partition, min_size = 3)
}
\arguments{
\item{partition}{- graph partition, a named vector.}

\item{min_size}{- size cut off.}
}
\value{
none
}
\description{
remove modules less than a minimum size from a partition
}
\examples{
filter_modules(partition, min_size = 5)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
