% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_proteins.R
\name{filter_proteins}
\alias{filter_proteins}
\title{filter_proteins}
\usage{
filter_proteins(data_in, colID)
}
\arguments{
\item{data_in}{- expression data}
}
\value{
none
}
\description{
remove one hit wonders and proteins that are not identified in at least
50% of samples.
}
\examples{
filter_proteins(data_in, colID)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
