% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_QC.R
\name{filter_QC}
\alias{filter_QC}
\title{filter_QC}
\usage{
filter_QC(data_in, groups, nbins, threshold)
}
\arguments{
\item{data_in}{- expression data}
}
\value{
none
}
\description{
This function bins peptides by mean intensity of QC samples, calculates
the log ratio of all QC sample comparisons and then removes a peptides
if its mean QC ratio is outside NxSD the mean its intensity bin.
}
\examples{
filter_QC(data_in, groups, nbins, threshold)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
