% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMedoid.R
\name{getMedoid}
\alias{getMedoid}
\title{getMedoid}
\usage{
getMedoid(adjm, k = NULL, h = NULL, method = "ward.D2")
}
\arguments{
\item{adjm}{adjacency matrix for heirarchical clustering}
}
\value{
none
}
\description{
Find representative branch from groups determined by heirarchical clustering
}
\examples{
getMedoid(adjm, h)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
