% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotBoxPlot.R
\name{ggplotBoxPlot}
\alias{ggplotBoxPlot}
\title{ggplotBoxPlot}
\usage{
ggplotBoxPlot(data_in, colID, colors, title)
}
\arguments{
\item{data_in}{- expression data.}
}
\value{
none
}
\description{
plot sample boxplots
}
\examples{
ggplotBoxPlot(data_in, colors, title)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
