% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotCorQC.R
\name{ggplotCorQC}
\alias{ggplotCorQC}
\title{ggplotCorQC}
\usage{
ggplotCorQC(data_in, groups, colID, nbins, annotate = TRUE)
}
\arguments{
\item{data_in}{- expression data}
}
\value{
none
}
\description{
Scatter plot showing correlation of QC samples. All comparisons.
nbin number of histograms showind distribution of QC ratios for all intensity bins.
}
\examples{
ggplotCorQC(data_in, groups, colID, nbins, annotate = TRUE)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
