% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotDensity.R
\name{ggplotDensity}
\alias{ggplotDensity}
\title{ggplotDensity}
\usage{
ggplotDensity(data_in, colID, title)
}
\arguments{
\item{data_in}{- expression data}
}
\value{
none
}
\description{
plot sample density plots
}
\examples{
ggplotDensity(data_in, title)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
