% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotDetect.R
\name{ggplotDetect}
\alias{ggplotDetect}
\title{ggplotDetect}
\usage{
ggplotDetect(data_in, group, log = TRUE)
}
\arguments{
\item{data_in}{- expression data.}
}
\value{
none
}
\description{
Function for plotting distribution (density) of missing values).
}
\examples{
ggplotDetect(data_in, group, log = TRUE)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
