% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotFreqOverlap.R
\name{ggplotFreqOverlap}
\alias{ggplotFreqOverlap}
\title{ggplotFreqOverlap}
\usage{
ggplotFreqOverlap(data_in, colID, groups)
}
\arguments{
\item{data_in}{- input data}
}
\value{
none
}
\description{
Function for plotting overlap frequency for peptides and proteins:
}
\examples{
ggplotFreqOverlap(data_in, colID, groups)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
