% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotGOscatter.R
\name{ggplotGOscatter}
\alias{ggplotGOscatter}
\title{ggplotGOscatter}
\usage{
ggplotGOscatter(GO_results, color, topN = 15)
}
\arguments{
\item{GO_results}{- df}
}
\value{
none
}
\description{
Function for visualizing GO terms.
}
\examples{
ggplotGOscatter(GO_results, color, topN = 10)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
