% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotMDS.R
\name{ggplotMDS}
\alias{ggplotMDS}
\title{ggplotMDS}
\usage{
ggplotMDS(data_in, colID, traits, title)
}
\arguments{
\item{data_in}{- expression data}
}
\value{
none
}
\description{
Generate MDS plot with ggplot.
Does not log transform the data.
}
\examples{
ggplotMDS(data_in, colID, traits, title)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
