% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotMeanSdPlot.R
\name{ggplotMeanSdPlot}
\alias{ggplotMeanSdPlot}
\title{ggplotMeanSdPlot}
\usage{
ggplotMeanSdPlot(data_in, colID, title, log)
}
\arguments{
\item{data_in}{- expression data}
}
\value{
none
}
\description{
wrapper around vsn::meanSdPlot that generates a mean Sd ggplot
}
\examples{
ggplotMeanSdPlot(data_in, colID, title, log)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
