% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotPCA.R
\name{ggplotPCA}
\alias{ggplotPCA}
\title{ggplotPCA}
\usage{
ggplotPCA(data_in, traits, colors, colID = "Abundance", title = "2D PCA Plot")
}
\arguments{
\item{data_in}{- the expression data}
}
\value{
none
}
\description{
plot a PCA plot with ggplot
}
\examples{
ggplotPCA(data_in, traits, colors)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
