% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotPeptideBarPlot.R
\name{ggplotPeptideBarPlot}
\alias{ggplotPeptideBarPlot}
\title{ggplotPeptideBarPlot}
\usage{
ggplotPeptideBarPlot(raw_peptide, seed = 1)
}
\arguments{
\item{raw_peptide}{- raw peptide expression data}
}
\value{
none
}
\description{
ggplotPeptideBarPlot
}
\examples{
ggplotPeptideBarPlot(raw_peptide)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
