% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotProteinBoxPlot.R
\name{ggplotProteinBoxPlot}
\alias{ggplotProteinBoxPlot}
\title{ggplotProteinBoxPlot}
\usage{
ggplotProteinBoxPlot(
  data_in,
  interesting.proteins,
  traits,
  box_order = NULL,
  scatter = FALSE
)
}
\arguments{
\item{data_in}{- expression data}
}
\value{
none
}
\description{
Function to make boxplots for proteins of interst.
}
\examples{
ggplotProteinBoxPlot(data_in, interesting.proteins, traits, scatter = TRUE)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
