% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotProteinScatterPlot.R
\name{ggplotProteinScatterPlot}
\alias{ggplotProteinScatterPlot}
\title{ggplotProteinScatterPlot}
\usage{
ggplotProteinScatterPlot(cleanDat, prot1, prot2)
}
\arguments{
\item{none}{}
}
\value{
None
}
\description{
Generate a scatter plot between two proteins.
}
\examples{
ggplotProteinScatterPlot(cleanDat, prot1, prot2)
}
\references{
\url{}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{}
