% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotQCHist.R
\name{ggplotQCHist}
\alias{ggplotQCHist}
\title{ggplotQCHist}
\usage{
ggplotQCHist(data_in, group, nbins, threshold)
}
\arguments{
\item{data_in}{- expression data}
}
\value{
none
}
\description{
plot histogram of qc ratios
}
\examples{
ggplotQCHist(data_in, groups, nbins, threshold)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
