% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotSampleConnectivity.R
\name{ggplotSampleConnectivity}
\alias{ggplotSampleConnectivity}
\title{ggplotSampleConnectivity}
\usage{
ggplotSampleConnectivity(data_in, log = TRUE, colID, threshold = -2.5)
}
\arguments{
\item{data_in}{- expression data}
}
\value{
none
}
\description{
Function for plotting sample connectivity.
}
\examples{
ggplotSampleConnectivity(data_in, log = TRUE, colID, threshold = -2.5)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
