% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotScaleFreeFit.R, R/ggplotScaleFreePlot.R
\name{ggplotScaleFreeFit}
\alias{ggplotScaleFreeFit}
\title{ggplotScaleFreeFit}
\usage{
ggplotScaleFreeFit(
  connectivity,
  nBreaks = 10,
  truncated = FALSE,
  removeFirst = FALSE,
  main = "",
  ...
)

ggplotScaleFreeFit(
  connectivity,
  nBreaks = 10,
  truncated = FALSE,
  removeFirst = FALSE,
  main = "",
  ...
)
}
\value{
none

none
}
\description{
evaluate the scale free fit of a graph

evaluate the scale free fit of a graph
}
\examples{
ggplotScaleFreePlot(connectivity)
ggplotScaleFreePlot(connectivity)
}
\references{
none

none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}

Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
