% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotVerboseBoxplot.R
\name{ggplotVerboseBoxplot}
\alias{ggplotVerboseBoxplot}
\title{ggplotVerboseBoxplot}
\usage{
ggplotVerboseBoxplot(x, groups, group_levels, control_group = NULL)
}
\arguments{
\item{x}{- ME vector}

\item{g}{- groups, same dimension as ME}

\item{contrasts}{- which groups to compare}

\item{order}{- order of the bars in the plot}
}
\value{
verbose boxplot
}
\description{
Generate WGCNA verbose boxplots
}
\examples{
ggplotVerboseBoxplot(x, g, contrasts)
}
\references{
\url{}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{}
