% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsavePlots.R
\name{ggsavePlots}
\alias{ggsavePlots}
\title{ggsavePlots}
\usage{
ggsavePlots(plots, filenames = NULL, ...)
}
\arguments{
\item{plots}{- list of plots}

\item{filenames}{- filenames for saving plots}
}
\value{
None
}
\description{
save a list of plots
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
