% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_peptides.R
\name{impute_peptides}
\alias{impute_peptides}
\title{impute_peptides}
\usage{
impute_peptides(data_in, groups, method, qc_threshold = 0, bio_threshold = 2)
}
\arguments{
\item{data_in}{- expression data}
}
\value{
none
}
\description{
Function to impute peptide level missing values.
Supports MLE for MAR and KNN for MNAR.
}
\examples{
impute_peptides(data_in, groups, method)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
