% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_proteins.R
\name{impute_proteins}
\alias{impute_proteins}
\title{impute_proteins}
\usage{
impute_proteins(data_in, colID, method = "knn")
}
\arguments{
\item{data_in}{- expression data}
}
\value{
none
}
\description{
A function for imputing TMT protein expression matrix with the KNN algorithm
from package impute. Note data is log transformed before imputing and then
return un-logged.
}
\examples{
impute_proteins(data_in)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
