% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePairwiseContrasts.R
\name{makePairwiseContrasts}
\alias{makePairwiseContrasts}
\title{makePairwiseContrasts}
\usage{
makePairwiseContrasts(g1, g2, collapse = " - ")
}
\arguments{
\item{g1}{- group1}

\item{g2}{- group2}
}
\value{
none
}
\description{
Function for making pairwise contrasts.
}
\examples{
makePairwiseContrasts(g1, g2, collapse = " - ")
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
