% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgi_batch_query.R
\name{mgi_batch_query}
\alias{mgi_batch_query}
\title{mgi_batch_query}
\usage{
mgi_batch_query(
  ids,
  from = "Uniprot",
  to = "Entrez",
  download = TRUE,
  quiet = TRUE
)
}
\description{
Maps Uniprot Accession identifiers to Entrez gene identifiers with MGI.
}
