% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named.list.R
\name{named.list}
\alias{named.list}
\title{named.list}
\usage{
named.list(...)
}
\value{
None
}
\description{
Create a named list.
}
\examples{
named.list()
}
\references{
\url{https://stackoverflow.com/questions/17831015/r-list-from-variables-using-variable-names}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{list}
\keyword{named}
