% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_IRS.R
\name{normalize_IRS}
\alias{normalize_IRS}
\title{normalize_IRS}
\usage{
normalize_IRS(data_in, IRS_ID, groups, robust = TRUE)
}
\arguments{
\item{data_in}{- expression data}
}
\value{
none
}
\description{
Performs IRS normalization
Supports geometric or robust mean.
}
\examples{
normalize_IRS(data_in,IRS_ID="QC",groups)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
