% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_SL.R
\name{normalize_SL}
\alias{normalize_SL}
\title{normalize_SL}
\usage{
normalize_SL(data_in, colID = "", group = "")
}
\arguments{
\item{data_in}{- expression data.}
}
\value{
none
}
\description{
Performs sample loading normalization based on column sums of TMT-runs.
This function also removes rows in which qc was not quantified in all three replicates.
}
\examples{
normalize_SL(data_in, colID = "", group = "")
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
