% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_TMM.R
\name{normalize_TMM}
\alias{normalize_TMM}
\title{normalize_TMM}
\usage{
normalize_TMM(data_in, groups)
}
\arguments{
\item{data_in}{- expression data}
}
\value{
none
}
\description{
perform tmm normalization using the edgeR package on the columns
specified by tmt_cols.
}
\examples{
normalize_TMM(data_in, colID)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
