% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntile.R
\name{ntile}
\alias{ntile}
\title{ntile}
\usage{
ntile(
  x,
  ngroups,
  na.rm = FALSE,
  result = "list",
  reverse = FALSE,
  checkBleed = TRUE
)
}
\arguments{
\item{x}{}

\item{ngroups}{}

\item{na.rm}{}

\item{result}{}

\item{reverse}{}

\item{checkBleed}{}
}
\value{
none
}
\description{
Create groups where the groups each have as close to the same number of
members as possible.
}
\examples{
ntile(setNames(state.area, state.name), 10)
}
\references{
\href{https://www.rdocumentation.org/packages/BurStMisc/versions/1.1/topics/ntile}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{bin}
\keyword{distribution}
\keyword{groups}
