% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_overlap.R
\name{peptide_overlap}
\alias{peptide_overlap}
\title{peptide_overlap}
\usage{
peptide_overlap(data_in, contrasts, info_cols)
}
\arguments{
\item{data_in}{- input data}

\item{contrasts}{- contrasts}

\item{info_cols}{- cols to be ignored.}
}
\value{
none
}
\description{
Calculates peptide identification overlap for given contrasts matrix.
}
\examples{
peptide_overlap(data_in, contrasts, info_cols)
}
\references{
\url{}
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{}
