% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_file.R
\name{prefix_file}
\alias{prefix_file}
\title{prefix_file
add a prefix to filename}
\usage{
prefix_file(
  file_path,
  file_name = NULL,
  output_dir = NULL,
  width = 2,
  date = FALSE
)
}
\description{
prefix_file
add a prefix to filename
}
