% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiet.R
\name{quiet}
\alias{quiet}
\title{quiet}
\usage{
quiet(x)
}
\arguments{
\item{x}{- expression to be silenced.}
}
\value{
none
}
\description{
Function for supressing printed messages from a function.
}
\examples{
quiet(myfun(x))
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
