% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rip.R
\name{rip}
\alias{rip}
\title{rip}
\usage{
rip(package, method = "utils", ...)
}
\arguments{
\item{package}{- the package to be downloaded}
}
\value{
none
}
\description{
convienence wrapper around several R installation methods
}
\examples{
rip("WGCNA")
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
