% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_free_fit.R
\name{scale_free_fit}
\alias{scale_free_fit}
\title{scale_free_fit}
\usage{
scale_free_fit(
  connectivity,
  nBreaks = 10,
  truncated = FALSE,
  removeFirst = FALSE,
  main = "",
  ...
)
}
\arguments{
\item{connectivity}{- degree centrality vector}
}
\value{
none
}
\description{
evaluate the scale free fit of a graph
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
