% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_protein.R
\name{summarize_protein}
\alias{summarize_protein}
\title{summarize_protein}
\usage{
summarize_protein(peptide_data)
}
\arguments{
\item{data_in}{- expression data.}
}
\value{
none
}
\description{
Summarize proteins by summing all peptides for a unique Accession identifier.
}
\examples{
summarize_protein(peptide_data)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
