% Generated by roxygen2: do not edit by hand
% Please edit documentation in TAMPOR/TAMPOR.R
\name{TAMPOR}
\alias{TAMPOR}
\title{TAMPOR}
\usage{
TAMPOR(dat, traits, noGIS = FALSE, useAllNonGIS = FALSE,
  batchPrefixInSampleNames = FALSE, GISchannels = "GIS",
  iterations = 250, samplesToIgnore = FALSE, meanOrMedian = "median",
  removeGISafter = FALSE, minimumBatchSize = 5, parallelThreads = 2,
  outputSuffix = "TAMPOR", path = getwd())
}
\arguments{
\item{dat}{}
}
\value{
none
}
\description{
performs TAMPOR normalization.
}
\examples{
TAMPOR(dat, traits)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
