% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotHistConnectivity.R
\name{ggplotHistConnectivity}
\alias{ggplotHistConnectivity}
\title{ggplotHistConnectivity}
\usage{
ggplotHistConnectivity(connectivity)
}
\value{
none
}
\description{
Network connectivity histogram.
}
\examples{
ggplotHistConnectivity(connectivity)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
